test CRUD Document in order

+++ tests/DocumentCRUDTest.php
@@ -4,9 +4,10 @@ namespace Mifiel\Tests;
 use Mifiel\ApiClient;
 use Mifiel\Document;
 
-class DocumentTest extends \PHPUnit_Framework_TestCase {
+class DocumentCRUDTest extends \PHPUnit_Framework_TestCase {
 
   const ORIGINAL_HASH = '7cf3c80bbe595734a960b49a79d6e87f8932c21fea6665b69cfa1257f85f7dc1';
+  private static $id;
 
   public function setTokens() {
     ApiClient::setTokens(
@@ -19,6 +20,18 @@ class DocumentTest extends \PHPUnit_Framework_TestCase {
     $this->setTokens();
     $documents = Document::all();
     return $documents[count($documents) - 1];
+  }
+
+  public function testSaveCreate() {
+    $this->setTokens();
+    $document = new Document([
+      'original_hash' => self::ORIGINAL_HASH,
+    ]);
+    $document->save();
+    self::$id = $document->id;
+    // Fetch document again
+    $document = $this->getDocument();
+    $this->assertEquals(self::ORIGINAL_HASH, $document->original_hash);
   }
 
   public function testSaveUpdate() {
@@ -41,8 +54,7 @@ class DocumentTest extends \PHPUnit_Framework_TestCase {
 
   public function testGetProperties() {
     $document = $this->getDocument();
-    $id = '708c8b9d-5440-46ef-8b67-5297aaf058d5';
-    $this->assertEquals($id, $document->id);
+    $this->assertEquals(self::$id, $document->id);
   }
 
   public function testSetProperties() {
@@ -56,7 +68,9 @@ class DocumentTest extends \PHPUnit_Framework_TestCase {
 
   public function testDelete() {
     $document = $this->getDocument();
-    $document->delete();
+    if ($document) {
+      $document->delete();
+    }
     $documents = Document::all();
     $this->assertEmpty($documents);
   }
